+++ gui/inventory/inventoryeditortab.cpp
@@ -10,7 +10,7 @@ InventoryEditorTab::InventoryEditorTab(const inventory_area* area, QWidget* pare
   ui->setupUi(this);
 
   slotViews.resize(area->count);
-  for (size_t i = 0; i < slotViews.size(); i++)
+  for (i64 i = 0; i < slotViews.size(); i++)
   {
     slotViews[i] = new ItemSlotView(area, i, this);
     slotViews[i]->setFixedSize(128, 128);
@@ -34,7 +34,7 @@ void InventoryEditorTab::Load(mhw_save_raw* mhwSave, int slotIndex)
 {
   SaveLoader::Load(mhwSave, slotIndex);
 
-  for (size_t i = 0; i < slotViews.count(); i++)
+  for (i64 i = 0; i < slotViews.count(); i++)
   {
     ItemSlotView* slotView = slotViews[i];
     slotView->Load(mhwSave, slotIndex);
@@ -57,7 +57,7 @@ void InventoryEditorTab::LoadIndex(int index)
 
 void InventoryEditorTab::LoadResources(ItemDB* itemDB, BitmapDB* bitmapDB)
 {
-  for (size_t i = 0; i < slotViews.count(); i++)
+  for (i64 i = 0; i < slotViews.count(); i++)
   {
     ItemSlotView* slotView = slotViews[i];
     slotView->LoadResources(itemDB, bitmapDB);
